Add validateEmail() to Field_Validation
validateEmail() is a helper method that checks that an email is in an
email format, returning "-1" if it is not, or the original email if it
is. This implementation is a *little* suspect because it won't accept
some emails, but every form of email format validation has tradeoffs and
this one is provided by Professor Garzon.