Add UserGettersTest.java
Added a JUnit test UserGettersTest.java. This was a real "a-ha!" moment.
Tests need to be placed in their own files following the provided
example in app.java.com.example.a2105projectgroup14.ExampleUnitTest.
This allows us to test our code in a standardized way that can be easily
run on the Android Development platform.